removed redirects from editOrder
